scripts/requirements.txt
requests==2.31.0
progressbar2==4.2.0
black
mysql-connector-python==8.2.0
unidecode==1.3.7